Minor renaming refactor
